+++ types/buttons.ts
@@ -1,0 +1,4 @@
+export interface IVideoPlayerButtons {
+  watchlist?: boolean;
+  favourite?: boolean;
+}
