+++ types/drm.ts
@@ -1,0 +1,7 @@
+export interface IVideoPlayerDRM {
+  id: number;
+  contentUrl: string;
+  drmScheme: string;
+  licensingServerUrl: string;
+  croToken: string;
+}
