+++ types/ima.ts
@@ -1,0 +1,7 @@
+export interface IVideoPlayerIMA {
+  adTagParameters?: Record<string, string>;
+  assetKey?: string;
+  authToken?: string;
+  contentSourceId?: string;
+  videoId?: string;
+}
