--- types/index.ts
@@ -1,4 +1,0 @@
-export * from './SeekToCommand';
-export * from './TextTrackTypes';
-export * from './VideoPlayer';
-export * from './VideoResizeMode';
