+++ types/metadata.ts
@@ -1,0 +1,7 @@
+export interface IVideoPlayerMetadata {
+  id: string;
+  description: string;
+  thumbnailUrl: string;
+  title: string;
+  type: string;
+}
