+++ types/mux.ts
@@ -1,0 +1,10 @@
+export interface IMuxData {
+  envKey: string;
+  subPropertyId: string;
+  viewerUserId: string;
+  videoId: string;
+  videoDuration: number;
+  videoIsLive: boolean;
+  videoStreamType: string;
+  videoTitle: string;
+}
