+++ types/resizeMode.ts
