+++ types/subtitles.ts
@@ -1,3 +1,9 @@
+export interface IVideoPlayerSubtitles {
+  language: string;
+  type: TextTrackType;
+  uri: string;
+}
+
 export enum TextTrackType {
   SRT = 'application/x-subrip',
   TTML = 'application/ttml+xml',
