+++ types/theme.ts
@@ -1,0 +1,11 @@
+
+export interface IVideoPlayerTheme {
+  colors: IVideoPlayerThemeObject,
+  fonts: IVideoPlayerThemeObject
+}
+
+
+interface IVideoPlayerThemeObject {
+  primary?: string;
+  secondary?: string;
+}
