+++ types/translations.ts
@@ -1,0 +1,11 @@
+interface IVideoPlayerTranslations {
+  player_epg_button: string;
+  player_stats_button: string;
+  player_play_button: string;
+  player_pause_button: string;
+  player_audio_and_subtitles_button: string;
+  // new labels
+  favourite: string;
+  watchlist: string;
+  moreVideos: string;
+}
